refactor(fallback): use OMP native retry chains, inline pi policy
Entire-Checkpoint: 532bf810626b
