+++ config/pi/default.nix
@@ -19,11 +19,6 @@ _: {
     source = ./fallback.json;
     force = true;
   };
-  # Kept outside extensions/ so the loader does not pick it up as an extension.
-  home.file.".pi/agent/fallback-policy.ts" = {
-    source = ../shared/fallback/runtime-fallback.ts;
-    force = true;
-  };
   home.file.".pi/agent/extensions/fallback.ts" = {
     source = ./fallback.ts;
     force = true;
